Add product offer listing functionality
FCO client is now capable of listing product offers that are
configured for selected user.

In order to get to this stage, we created base resource that all other
resources will be derived from. This makes it possible to keep common
functionality in single class.

We as implemented actual REST client that talks to remote service.
It can communicate with any FCO endpoint, but is currently configured
onyl for product offers. This will change as we gradually add more
functionality to the client.

In order to make this functionality usable, we created command line
client named fco that can be used to display offers from the comfort
of our favourite terminal emulator.