Remove unsafe default value
Having empty dictionary as a default parameter value is just calling
for trouble, since this dictionary will only be initiated once and
then reused on all function calls. Combine this with dictionary's
mutability and we have some seriously hard-to-debug bugs coming our
way.

Instead of having empty dict as default value, we used None to mark
the missing parameter and handle this using some conditional logic.