Sort json keys on output
Having output of the command non-deterministically permuted on each
invocation can be really frustrating. In order to make outputs of json
serialization deterministic, we force keys to be sorted.

Comparing outputs that are sorted using diff tool is far more useful,
since there is no moving of keys, just deletions and insertions. And
as an additional benefit, searching through output for certain key can
now (in theory) be done in O(log n) time;).