Add job client and data types
Job is central object of the asynchronous part of the API. These
objects are used to monitor the operation progress and termination.

Command line client is capable of listing jobs that match user
specified filter. This filtering is indispensable, since it is
practically impossible to fetch all of the jobs.